add: platform configuration
